+++ packages/eyes-rendering/lib/Eyes.js
@@ -2,8 +2,9 @@
 
 const { makeVisualGridClient } = require('@applitools/visual-grid-client');
 const { getProcessPageAndSerializeScript } = require('@applitools/dom-capture');
-const { Logger, ArgumentGuard, TypeUtils } = require('@applitools/eyes-common');
-const { BatchInfo, TestResultsFormatter } = require('@applitools/eyes-sdk-core');
+const { Logger, ArgumentGuard, Configuration, TypeUtils } = require('@applitools/eyes-common');
+const { BatchInfo, RectangleSize, TestFailedError, TestResultsFormatter } = require('@applitools/eyes-sdk-core');
+const { EyesSeleniumUtils, EyesWebDriver } = require('@applitools/eyes-selenium');
 const { RenderingConfiguration } = require('./RenderingConfiguration');
 
 class Eyes {
@@ -14,12 +15,16 @@ class Eyes {
     /** @type {EyesJsExecutor} */ this._jsExecutor = undefined;
     /** @type {ProxySettings} */ this._proxy = undefined;
     /** @type {BatchInfo} */ this._batchInfo = new BatchInfo();
+    /** @type {Configuration} */ this._configuration = new Configuration();
+    /** @type {boolean} */ this._sendDom = true;
 
     this._isOpen = false;
     this._processPageAndSerializeScript = undefined;
 
     this._checkWindowCommand = undefined;
     this._closeCommand = undefined;
+
+    /** @type {boolean} */ this._isVisualGrid = true;
   }
 
   /**
@@ -34,9 +39,12 @@ class Eyes {
 
   /**
    * @param {WebDriver} webDriver
+   * @param appName
+   * @param testName
+   * @param viewportSize
    * @param {RenderingConfiguration} renderingConfiguration
    */
-  async open(webDriver, _renderingConfiguration) {
+  async open(webDriver, appName, testName, viewportSize, _renderingConfiguration) {
     this._logger.verbose('enter');
 
     ArgumentGuard.notNull(webDriver, 'webDriver');
@@ -44,14 +52,21 @@ class Eyes {
 
     const renderingConfiguration = (_renderingConfiguration instanceof RenderingConfiguration) ? _renderingConfiguration :
       RenderingConfiguration.fromObject(_renderingConfiguration);
-    const apiKey = process.env.APPLITOOLS_API_KEY;
+    const apiKey = this.getApiKey();
     const showLogs = process.env.APPLITOOLS_SHOW_LOGS;
     const saveDebugData = process.env.APPLITOOLS_SAVE_DEBUG_DATA;
 
     await this._initDriver(webDriver);
 
-    if (renderingConfiguration.getBatch() == null) {
-      renderingConfiguration.setBatch(this._batchInfo);
+    if (!appName) {
+      appName = renderingConfiguration.getAppName() ? renderingConfiguration.getAppName() : this.getAppName();
+    } else {
+      this.setAppName(appName);
+    }
+    if (!testName) {
+      testName = renderingConfiguration.getTestName() ? renderingConfiguration.getTestName() : this.getTestName();
+    } else {
+      this.setTestName(testName);
     }
 
     const { openEyes } = makeVisualGridClient({
@@ -67,28 +82,33 @@ class Eyes {
 
     this._logger.verbose('opening openEyes...');
 
+    viewportSize = viewportSize ? viewportSize : this.getViewportSize();
+    if (viewportSize) {
+      await this.setViewportSize(viewportSize);
+    }
+
     const { checkWindow, close } = await openEyes({
-      appName: renderingConfiguration.getAppName(),
-      testName: renderingConfiguration.getTestName(),
+      appName: appName,
+      testName: testName,
       browser: renderingConfiguration.getBrowsersInfo(),
 
       // properties,
-      batchName: renderingConfiguration.getBatch().getName(),
-      batchId: renderingConfiguration.getBatch().getId(),
-      baselineBranchName: renderingConfiguration.getBaselineBranchName(),
-      baselineEnvName: renderingConfiguration.getBaselineEnvName(),
-      // baselineName,
-      envName: renderingConfiguration.getEnvironmentName(),
-      // ignoreCaret,
-      // isDisabled,
-      // matchLevel,
-      // matchTimeout,
-      parentBranchName: renderingConfiguration.getParentBranchName(),
-      branchName: renderingConfiguration.getBranchName(),
-      // saveFailedTests,
-      // saveNewTests,
-      // compareWithParentBranch,
-      // ignoreBaseline,
+      batchName: renderingConfiguration.getBatch() && renderingConfiguration.getBatch().getName() ? renderingConfiguration.getBatch().getName() : this.getBatch().getName(),
+      batchId: renderingConfiguration.getBatch() && renderingConfiguration.getBatch().getId() ? renderingConfiguration.getBatch().getId() : this.getBatch().getId(),
+      baselineBranchName: renderingConfiguration.getBaselineBranchName() ? renderingConfiguration.getBaselineBranchName() : this.getBaselineBranchName(),
+      baselineEnvName: renderingConfiguration.getBaselineEnvName() ? renderingConfiguration.getBaselineEnvName() : this.getBaselineEnvName(),
+      baselineName: renderingConfiguration.getBaselineEnvName() ? renderingConfiguration.getBaselineEnvName() : this.getBaselineEnvName(),
+      envName: renderingConfiguration.getEnvironmentName() ? renderingConfiguration.getEnvironmentName() : this.getEnvironmentName(),
+      ignoreCaret: this.getIgnoreCaret(),
+      isDisabled: this.isDisabled(),
+      matchLevel: this.getMatchLevel(),
+      matchTimeout: this.getMatchTimeout(),
+      parentBranchName: renderingConfiguration.getParentBranchName() ? renderingConfiguration.getParentBranchName() : this.getParentBranchName(),
+      branchName: renderingConfiguration.getBranchName() ? renderingConfiguration.getBranchName() : this.getBranchName(),
+      saveFailedTests: this.getSaveFailedTests(),
+      saveNewTests: this.getSaveNewTests(),
+      compareWithParentBranch: this.isCompareWithParentBranch(),
+      ignoreBaseline: this.isIgnoreBaseline(),
       logger: this._logger,
       // renderBatch,
       // waitForRenderedStatus,
@@ -98,7 +118,7 @@ class Eyes {
       // getRenderInfo,
       // createRGridDOMAndGetResourceMapping,
       // eyesTransactionThroat,
-      agentId: renderingConfiguration.getAgentId(),
+      agentId: renderingConfiguration.getAgentId() ? renderingConfiguration.getAgentId() : this.getAgentId()
     });
 
     this._checkWindowCommand = checkWindow;
@@ -115,6 +135,12 @@ class Eyes {
     if (TypeUtils.hasMethod(webDriver, ['executeScript', 'executeAsyncScript'])) {
       this._jsExecutor = webDriver;
     }
+
+    if (webDriver instanceof EyesWebDriver) {
+      this._driver = webDriver;
+    } else {
+      this._driver = new EyesWebDriver(this._logger, this, webDriver);
+    }
   }
 
   /**
@@ -124,7 +150,8 @@ class Eyes {
    * @return {Promise<TestResults[]>}
    */
   async close(throwEx) {
-    return this.closeAndReturnResults(throwEx);
+     const results = await this.closeAndReturnResults(throwEx);
+     return results && results.length > 0 ? results[0] : null;
   }
 
   // noinspection JSMethodCanBeStatic
@@ -150,12 +177,23 @@ class Eyes {
     return false;
   }
 
-  // noinspection JSMethodCanBeStatic
   /**
-   * @return {boolean}
+   * Sets the API key of your applitools Eyes account.
+   *
+   * @param apiKey {string} The api key to be used.
+   */
+  setApiKey(apiKey) {
+    ArgumentGuard.notNull(apiKey, 'apiKey');
+    ArgumentGuard.alphanumeric(apiKey, 'apiKey');
+    this._configuration.setApiKey(apiKey);
+  }
+
+  // noinspection JSUnusedGlobalSymbols
+  /**
+   * @return {string} The currently set API key or {@code null} if no key is set.
    */
   getApiKey() {
-    return undefined; // TODO: should use configuration
+    return this._configuration.getApiKey();
   }
 
   /**
@@ -187,6 +225,13 @@ class Eyes {
    */
   setBatch(batchInfo) {
     this._batchInfo = batchInfo;
+  }
+
+  /**
+   * @return {BatchInfo} The currently set batch info.
+   */
+  getBatch() {
+    return this._batchInfo;
   }
 
   /**
@@ -232,6 +277,11 @@ class Eyes {
 
     this._logger.verbose(`Dom extraction starting   (${checkSettings.toString()})   $$$$$$$$$$$$`);
 
+    let targetSelector = await checkSettings.getTargetProvider();
+    if (targetSelector) {
+      targetSelector = await targetSelector.getSelector(this);
+    }
+
     const domCaptureScript = `var callback = arguments[arguments.length - 1]; return (${this._processPageAndSerializeScript})().then(JSON.stringify).then(callback, function(err) {callback(err.stack || err.toString())})`;
     const results = await this._jsExecutor.executeAsyncScript(domCaptureScript);
     const { cdt, url: pageUrl, blobs, resourceUrls } = JSON.parse(results);
@@ -252,14 +302,305 @@ class Eyes {
       cdt,
       tag: checkSettings.getName(),
       sizeMode: checkSettings.getSizeMode(),
-      selector: checkSettings.getSelector(),
-      region: checkSettings.getRegion(),
+      selector: targetSelector,
+      region: checkSettings.getTargetRegion(),
       scriptHooks: checkSettings.getScriptHooks(),
       ignore: checkSettings.getIgnoreRegions(),
       floating: checkSettings.getFloatingRegions(),
-      sendDom: checkSettings.getSendDom(),
+      sendDom: checkSettings.getSendDom() ? checkSettings.getSendDom() : this.getSendDom(),
+      matchLevel: checkSettings.getMatchLevel() ? checkSettings.getMatchLevel() : this.getMatchLevel()
     });
   }
+
+  /**
+   * @return {string}
+   */
+  getTestName() {
+    return this._configuration.getTestName();
+  }
+
+  /**
+   * @param {string} testName
+   */
+  setTestName(testName) {
+    this._configuration.setTestName(testName);
+  }
+
+  /**
+   * @param appName {string} The name of the application under test.
+   */
+  setAppName(appName) {
+    this._configuration.setAppName(appName);
+  }
+
+  // noinspection JSUnusedGlobalSymbols
+  /**
+   * @return {string} The name of the application under test.
+   */
+  getAppName() {
+    return this._configuration.getAppName();
+  }
+
+  /**
+   * @return {string}
+   */
+  getBaselineBranchName() {
+    return this._configuration.getBaselineBranchName();
+  }
+
+  /**
+   * @param {string} baselineBranchName
+   */
+  setBaselineBranchName(baselineBranchName) {
+    this._configuration.setBaselineBranchName(baselineBranchName);
+  }
+
+  /**
+   * @return {string}
+   */
+  getBaselineEnvName() {
+    return this._configuration.getBaselineEnvName();
+  }
+
+  /**
+   * @param {string} baselineEnvName
+   */
+  setBaselineEnvName(baselineEnvName) {
+    this._configuration.setBaselineEnvName(baselineEnvName);
+  }
+
+  /**
+   * @return {string}
+   */
+  getEnvironmentName() {
+    return this._configuration.getEnvironmentName();
+  }
+
+  /**
+   * @param {string} environmentName
+   */
+  setEnvironmentName(environmentName) {
+    this._configuration.setEnvironmentName(environmentName);
+  }
+
+  /**
+   * @param agentId
+   */
+  setAgentId(agentId) {
+    this._configuration.setAgentId(agentId);
+  }
+
+  /**
+   * @return {string}
+   */
+  getAgentId() {
+    return this._configuration.getAgentId();
+  }
+
+  /**
+   * @param parentBranchName
+   */
+  setParentBranchName(parentBranchName) {
+    this._configuration.setParentBranchName(parentBranchName);
+  }
+
+  /**
+   * @return {string}
+   */
+  getParentBranchName() {
+    return this._configuration.getParentBranchName();
+  }
+
+  /**
+   * @param branchName
+   */
+  setBranchName(branchName) {
+    this._configuration.setBranchName(branchName);
+  }
+
+  /**
+   * @return {string}
+   */
+  getBranchName() {
+    return this._configuration.getBranchName();
+  }
+
+  /**
+   * @param {boolean} compareWithParentBranch New compareWithParentBranch value, default is false
+   */
+  setCompareWithParentBranch(compareWithParentBranch) {
+    this._configuration.setCompareWithParentBranch(compareWithParentBranch);
+  }
+
+  /**
+   * @return {boolean} The currently compareWithParentBranch value
+   */
+  isCompareWithParentBranch() {
+    return this._configuration.isCompareWithParentBranch();
+  }
+
+  /**
+   * @param {boolean} ignoreBaseline New ignoreBaseline value, default is false
+   */
+  setIgnoreBaseline(ignoreBaseline) {
+    this._configuration.setIgnoreBaseline(ignoreBaseline);
+  }
+
+  /**
+   * @return {boolean} The currently ignoreBaseline value
+   */
+  isIgnoreBaseline() {
+    return this._configuration.isIgnoreBaseline();
+  }
+
+  /**
+   * Set whether or not new tests are saved by default.
+   *
+   * @param {boolean} saveNewTests True if new tests should be saved by default. False otherwise.
+   */
+  setSaveNewTests(saveNewTests) {
+    this._configuration.setSaveNewTests(saveNewTests);
+  }
+
+  /**
+   * @return {boolean} True if new tests are saved by default.
+   */
+  getSaveNewTests() {
+    return this._configuration.getSaveNewTests();
+  }
+
+  /**
+   * Set whether or not failed tests are saved by default.
+   *
+   * @param {boolean} saveFailedTests True if failed tests should be saved by default, false otherwise.
+   */
+  setSaveFailedTests(saveFailedTests) {
+    this._configuration.setSaveFailedTests(saveFailedTests);
+  }
+
+  // noinspection JSUnusedGlobalSymbols
+  /**
+   * @return {boolean} True if failed tests are saved by default.
+   */
+  getSaveFailedTests() {
+    return this._configuration.getSaveFailedTests();
+  }
+
+  /**
+   * Sets the maximum time (in ms) a match operation tries to perform a match.
+   * @param {number} ms Total number of ms to wait for a match.
+   */
+  setMatchTimeout(ms) {
+    this._configuration.setMatchTimeout(ms);
+  }
+
+  /**
+   * @return {number} The maximum time in ms {@link #checkWindowBase(RegionProvider, string, boolean, number)} waits
+   *   for a match.
+   */
+  getMatchTimeout() {
+    return this._configuration.getMatchTimeout();
+  }
+
+  /**
+   * The test-wide match level to use when checking application screenshot with the expected output.
+   *
+   * @param {MatchLevel} matchLevel The test-wide match level to use when checking application screenshot with the
+   *   expected output.
+   */
+  setMatchLevel(matchLevel) {
+    this._configuration.setMatchLevel(matchLevel);
+  }
+
+  /**
+   * @return {MatchLevel} The test-wide match level.
+   */
+  getMatchLevel() {
+    return this._configuration.getMatchLevel();
+  }
+
+  /**
+   * Sets the ignore blinking caret value.
+   *
+   * @param {boolean} value The ignore value.
+   */
+  setIgnoreCaret(value) {
+    this._configuration.setIgnoreCaret(value);
+  }
+
+  /**
+   * @return {boolean} Whether to ignore or the blinking caret or not when comparing images.
+   */
+  getIgnoreCaret() {
+    return this._configuration.getIgnoreCaret();
+  }
+
+  /**
+   * @param isDisabled {boolean} If true, all interactions with this API will be silently ignored.
+   */
+  setIsDisabled(isDisabled) {
+    this._configuration.setIsDisabled(isDisabled);
+  }
+
+  /**
+   * @return {boolean} Whether eyes is disabled.
+   */
+  isDisabled() {
+    return this._configuration.getIsDisabled();
+  }
+
+  /**
+   * @param {boolean} sendDom
+   */
+  setSendDom(sendDom) {
+    this._sendDom = sendDom;
+  }
+
+  // noinspection JSUnusedGlobalSymbols
+  /**
+   * @return {boolean}
+   */
+  getSendDom() {
+    return this._sendDom;
+  }
+
+  /**
+   * @return {RectangleSize}
+   */
+  getViewportSize() {
+    return this._configuration.getViewportSize();
+  }
+
+  /**
+   * @param {RectangleSize} viewportSize
+   */
+  async setViewportSize(viewportSize) {
+    ArgumentGuard.notNull(viewportSize, 'viewportSize');
+    viewportSize = new RectangleSize(viewportSize);
+    this._configuration.setViewportSize(viewportSize);
+
+    if (this._driver) {
+      const originalFrame = this._driver.getFrameChain();
+      await this._driver.switchTo().defaultContent();
+
+      try {
+        await EyesSeleniumUtils.setViewportSize(this._logger, this._driver, viewportSize);
+      } catch (err) {
+        await this._driver.switchTo().frames(originalFrame); // Just in case the user catches that error
+        throw new TestFailedError('Failed to set the viewport size', err);
+      }
+
+      await this._driver.switchTo().frames(originalFrame);
+    }
+  }
+
+  /**
+   * @return {boolean}
+   */
+  isVisualGrid() {
+    return this._isVisualGrid;
+  }
+
 }
 
 exports.Eyes = Eyes;