+++ packages/eyes-images-browser/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.5.3](https://github.com/applitools/eyes.sdk.javascript1/compare/@applitools/eyes-images-browser@4.5.2...@applitools/eyes-images-browser@4.5.3) (2019-01-22)
+
+**Note:** Version bump only for package @applitools/eyes-images-browser
+
+
+
+
+
 ## [4.5.2](https://github.com/applitools/eyes.sdk.javascript1/compare/@applitools/eyes-images-browser@4.5.1...@applitools/eyes-images-browser@4.5.2) (2019-01-21)
 
 **Note:** Version bump only for package @applitools/eyes-images-browser