+++ packages/eyes-rendering/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0](https://github.com/applitools/eyes.sdk.javascript1/compare/@applitools/eyes-rendering@1.1.3...@applitools/eyes-rendering@2.0.0) (2019-01-22)
+
+
+### feature
+
+* **eyes-rendering:** add ability to Target.region(By|WebElement), etc ([54c3a45](https://github.com/applitools/eyes.sdk.javascript1/commit/54c3a45))
+
+
+### BREAKING CHANGES
+
+* **eyes-rendering:** `Target.selector(string)` is removed, use `Target.region(string)` instead
+
+
+
+
+
 ## [1.1.3](https://github.com/applitools/eyes.sdk.javascript1/compare/@applitools/eyes-rendering@1.1.2...@applitools/eyes-rendering@1.1.3) (2019-01-21)
 
 **Note:** Version bump only for package @applitools/eyes-rendering